+++ routes/admin/index.ts
@@ -6,23 +6,22 @@ const secret = 'tadatada';
 
 router.use((req, res, next) => {
   const { originalUrl } = req;
-
   if (originalUrl.includes('/image')) return next();
 
-  if (originalUrl === '/admin/' + secret) {
+  if (originalUrl === '/next/admin/' + secret) {
     res.cookie('secret', secret, { maxAge: 3600000 }); // 1 hour
-    return res.redirect('/admin');
+    return res.redirect('/next/admin');
   }
 
   const s = req.cookies.secret;
 
   if (s && s === secret) return next();
 
-  res.redirect('/');
+  res.redirect('/next');
 });
 
 router.get('/', (req, res) => {
-  res.render('admin');
+  res.render('next/admin');
 });
 
 router.use('/schematic_changes', changesRouter);
