+++ routes/index.ts
@@ -1,7 +1,5 @@
 export { default as admin } from './admin';
 export { default as api } from './api';
 export { default as raw } from './raw';
-export { default as schematics } from './schematics';
-export { default as shorturl } from './shorturl';
-export { default as tutorials } from './tutorials';
-export { default as next } from './next';
+export { default as main } from './main';
+export { default as help } from './help';
