+++ routes/info.ts
@@ -1,0 +1,11 @@
+import { Router } from 'express';
+
+const router = Router();
+export default router;
+
+router.get('/info', (req, res) => {
+  res.render('next/info.pug');
+});
+router.get('/info/credits', (req, res) => {
+  res.render('next/credits.pug');
+});
